Second commit

+++ app/models/CalculationResults/InDatesTaxYearsCalculation.scala
@@ -26,9 +26,9 @@ case class InDatesTaxYearsCalculation(
   debit: Double,
   chargePaidByMember: Double,
   chargePaidBySchemes: Double,
-  revisedChargableAmountBeforeTaxRate: Double,
-  revisedChargableAmountAfterTaxRate: Double,
-  unusedAnnualAllowance: Double,
+  revisedChargableAmountBeforeTaxRate: Int,
+  revisedChargableAmountAfterTaxRate: Int,
+  unusedAnnualAllowance: Int,
   taxYearSchemes: List[InDatesTaxYearSchemeCalculation],
   totalCompensation: Option[Double]
 )
@@ -46,9 +46,9 @@ object InDatesTaxYearsCalculation {
         (__ \ "debit").read[Double] and
         (__ \ "chargePaidByMember").read[Double] and
         (__ \ "chargePaidBySchemes").read[Double] and
-        (__ \ "revisedChargableAmountBeforeTaxRate").read[Double] and
-        (__ \ "revisedChargableAmountAfterTaxRate").read[Double] and
-        (__ \ "unusedAnnualAllowance").read[Double] and
+        (__ \ "revisedChargableAmountBeforeTaxRate").read[Int] and
+        (__ \ "revisedChargableAmountAfterTaxRate").read[Int] and
+        (__ \ "unusedAnnualAllowance").read[Int] and
         (__ \ "taxYearSchemes").read[List[InDatesTaxYearSchemeCalculation]] and
         (__ \ "totalCompensation").readNullable[Double]
     )(InDatesTaxYearsCalculation.apply _)
@@ -65,9 +65,9 @@ object InDatesTaxYearsCalculation {
         (__ \ "debit").write[Double] and
         (__ \ "chargePaidByMember").write[Double] and
         (__ \ "chargePaidBySchemes").write[Double] and
-        (__ \ "revisedChargableAmountBeforeTaxRate").write[Double] and
-        (__ \ "revisedChargableAmountAfterTaxRate").write[Double] and
-        (__ \ "unusedAnnualAllowance").write[Double] and
+        (__ \ "revisedChargableAmountBeforeTaxRate").write[Int] and
+        (__ \ "revisedChargableAmountAfterTaxRate").write[Int] and
+        (__ \ "unusedAnnualAllowance").write[Int] and
         (__ \ "taxYearSchemes").write[List[InDatesTaxYearSchemeCalculation]] and
         (__ \ "totalCompensation").writeNullable[Double]
     )(a =>
